Develop #3.0.4 - before 3.1

+++ pod/video/views.py
@@ -28,7 +28,7 @@ from dateutil.parser import parse
 import concurrent.futures as futures
 
 from pod.main.views import in_maintenance
-from pod.main.decorators import ajax_required, admin_required
+from pod.main.decorators import ajax_required, ajax_login_required, admin_required
 from pod.authentication.utils import get_owners as auth_get_owners
 from pod.video.utils import get_videos as video_get_videos
 from pod.video.models import Video
@@ -89,10 +89,10 @@ TEMPLATE_VISIBLE_SETTINGS = getattr(
         "TITLE_ETB": "University name",
         "LOGO_SITE": "img/logoPod.svg",
         "LOGO_ETB": "img/logo_etb.svg",
-        "LOGO_PLAYER": "img/logoPod.svg",
+        "LOGO_PLAYER": "img/pod_favicon.svg",
         "LINK_PLAYER": "",
         "FOOTER_TEXT": ("",),
-        "FAVICON": "img/logoPod.svg",
+        "FAVICON": "img/pod_favicon.svg",
         "CSS_OVERRIDE": "",
         "PRE_HEADER_TEMPLATE": "",
         "POST_FOOTER_TEMPLATE": "",
@@ -1774,7 +1774,7 @@ def video_oembed(request):
     if not request.GET.get("url"):
         raise SuspiciousOperation("URL must be specified")
     format = "xml" if request.GET.get("format") == "xml" else "json"
-
+    version = request.GET.get("version") if request.GET.get("version") else "video"
     data = {}
     data["type"] = "video"
     data["version"] = "1.0"
@@ -1795,7 +1795,7 @@ def video_oembed(request):
     p = re.compile(reg)
     m = p.match(url)
 
-    if m:
+    if m and m.group("slug") not in ["add", "edit"]:
         video_slug = m.group("slug")
         slug_private = m.group("slug_private")
         try:
@@ -1811,6 +1811,14 @@ def video_oembed(request):
             reverse("videos:videos"),
             video.owner.username,
         )
+        video_url = (
+            reverse("video:video", kwargs={"slug": video.slug})
+            if version == "video"
+            else reverse(
+                "%(version)s:video_%(version)s" % {"version": version},
+                kwargs={"slug": video.slug},
+            )
+        )
         data["html"] = (
             '<iframe src="%(provider)s%(video_url)s%(slug_private)s'
             + '?is_iframe=true" width="640" height="360" '
@@ -1818,7 +1826,7 @@ def video_oembed(request):
             + "allowfullscreen loading='lazy'></iframe>"
         ) % {
             "provider": data["provider_url"],
-            "video_url": reverse("video:video", kwargs={"slug": video.slug}),
+            "video_url": video_url,
             "slug_private": "%s/" % slug_private if slug_private else "",
         }
     else:
@@ -2079,10 +2087,15 @@ def vote_get(request, video_slug):
         )
 
 
-@login_required(redirect_field_name="referrer")
+@ajax_login_required
+@csrf_protect
 def vote_post(request, video_slug, comment_id):
     if request.method == "GET":
         return HttpResponseNotFound("<h1>Method Not Allowed</h1>", status=405)
+    if in_maintenance():
+        return HttpResponseForbidden(
+            _("Sorry, you can't vote a comment while the server is under maintenance.")
+        )
     # current video
     c_video = get_object_or_404(Video, slug=video_slug)
     # current comment
@@ -2107,9 +2120,13 @@ def vote_post(request, video_slug, comment_id):
     return HttpResponse(json.dumps(response), content_type="application/json")
 
 
-@login_required(redirect_field_name="referrer")
+@ajax_login_required
 @csrf_protect
 def add_comment(request, video_slug, comment_id=None):
+    if in_maintenance():
+        return HttpResponseForbidden(
+            _("Sorry, you can't comment while the server is under maintenance.")
+        )
     if request.method == "POST":
         # current video
         c_video = get_object_or_404(Video, slug=video_slug)
@@ -2281,9 +2298,18 @@ def get_comments(request, video_slug):
         )
 
 
-@login_required(redirect_field_name="referrer")
+@ajax_login_required
+@csrf_protect
 def delete_comment(request, video_slug, comment_id):
+    """Delete the comment `comment_id` associated to `video_slug`.
 
+    Args:
+        video_slug (string): the video associated to this comment
+        comment_id (): id of the comment to be deleted
+
+    Returns:
+        HttpResponse
+    """
     v = get_object_or_404(Video, slug=video_slug)
     c_user = request.user
     c = get_object_or_404(Comment, video=v, id=comment_id)
@@ -2291,13 +2317,16 @@ def delete_comment(request, video_slug, comment_id):
         "deleted": True,
     }
 
-    if c.author == c_user or v.owner == c_user or c_user.is_superuser:
+    if in_maintenance():
+        return HttpResponseForbidden(
+            _("Sorry, you can't delete a comment while the server is under maintenance.")
+        )
 
+    if c.author == c_user or v.owner == c_user or c_user.is_superuser:
         c.delete()
         response["comment_deleted"] = comment_id
         return HttpResponse(json.dumps(response), content_type="application/json")
     else:
-
         response["deleted"] = False
         response["message"] = _("You do not have rights to delete this comment")
         return HttpResponse(
